DROTH-3295 revert extractGeometry location change and renaming
